chore(icons): ✨ add color control to wm-icon component
Updated the `wm-icon` component to allow control over whether the icon color is applied. This was achieved by introducing a new input property `useColor`, which defaults to `true`. If set to `false`, the icon will not be colored according to `iconData.color`.

- Modified `select-filter.component.html` to pass `[useColor]="false"` to `wm-icon`.
- Adjusted `status-filter.component.html` to use `wm-icon` instead of `appBuildSvg` for better consistency.
- Updated `filters.component.scss` to refine margin styling for `ion-label`.

These changes enhance flexibility in how icons are displayed across different components, enabling scenarios where color customization is not desired.