refactor(geolocation): 🔄 start navigation and update track recording
Added a call to `startNavigation` in the `resumeRecordingFromSaved` method to ensure navigation begins when resuming recording. Updated the store dispatch to set the current UGC track recording with the saved feature and time.

Removed unnecessary `setEnableTrackRecorderPanel$` effect from `user-activity.effects.ts` as it was redundant and the logic is now handled directly within the `GeolocationService`. This streamlines the process and reduces unnecessary code.