fix(geolocation): 🐛 remove pause recording call and update UGC track recording state
Removed the call to `pauseRecording` in `GeolocationService` as it was unnecessary and could cause issues with the recording state. Additionally, ensured that the UGC track recording state is reset properly by dispatching `setCurrentUgcTrackRecording` with default values.

Also, removed the call to `removeCurrentUgcTrack` in `localForage.ts` to prevent unintended removal of UGC track data after saving.